pyproject.toml
[tool.poetry]
name = "data-scientist-take-home"
version = "0.1.0"
description = ""
authors = ["Jayant Krishnamurthy <jayantkrishnamurthy@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
web3 = "^6.15.1"
eth-ape = "^0.7.8"
matplotlib = "^3.8.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
